move the classification print function to the data package
