+++ blackdoc/tests/test_classification.py
@@ -3,33 +3,7 @@ import more_itertools
 from blackdoc import classification
 
 from .data import doctest as data
-
-
-def print_line_with_range(name, range_, unit):
-    min_, max_ = range_
-    line_numbers = range(min_, max_)
-
-    no_group = " "
-    start_group = "┐"
-    mid_group = "│"
-    end_group = "┘"
-
-    for index, (lineno, line) in enumerate(zip(line_numbers, unit.split("\n"))):
-        if max_ - min_ == 1:
-            classifier = no_group
-        elif index == 0:
-            classifier = start_group
-        elif index == max_ - min_ - 1:
-            classifier = end_group
-        else:
-            classifier = mid_group
-
-        print(f"{name:>8s} {classifier} → {index:02d}: {line}")
-
-
-def print_classification(labeled):
-    for range, name, unit in labeled:
-        print_line_with_range(name, range, unit)
+from .data import print_classification
 
 
 def test_detect_format():
